requirements.txt
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.5.2
sqlalchemy==2.0.23
mistralai==0.0.12
python-multipart==0.0.6
jinja2==3.1.2
